Refactor buildParameters
