+++ lib/cfn.js
@@ -6,6 +6,7 @@ const roles = require('./resources/roles');
 const alarms = require('./resources/alarms');
 const dispatchResource = require('./resources/dispatch');
 const lambdaResource = require('./resources/lambda');
+const cloudfrontParameters = require('./parameters');
 
 function buildFunctionTemplate(options) {
   let template = buildFunction(options);
@@ -20,7 +21,7 @@ function buildFunction(options) {
     throw new Error('Function name is required');
   }
 
-  let parameters = buildParameters(options);
+  let parameters = cloudfrontParameters.buildParameters(options);
   let defaultAlarms = alarms.buildServiceAlarms(options);
   let lambda = lambdaResource.buildLambda(options);
   let role = roles.buildRole(options);
@@ -284,46 +285,6 @@ function compileFunction() {
   return template;
 }
 
-/* eslint-enable no-loop-func */
-function buildParameters(options) {
-  let parameters = {
-    Parameters: {}
-  };
-
-  if (options.parameters) {
-    for (let param in options.parameters) {
-      let isAlphaNumeric = /^[a-zA-Z0-9]+$/.test(param);
-      if (!isAlphaNumeric) {
-        throw new Error('Parameter names must be alphanumeric');
-      }
-
-      if (!options.parameters[param].Type) {
-        throw new Error('Parameter must contain Type property');
-      }
-
-      if (!options.parameters[param].Description) {
-        throw new Error('Parameter must contain Description property');
-      }
-    }
-    parameters.Parameters = options.parameters;
-  }
-
-  parameters.Parameters.CodeS3Bucket = {
-    Type: 'String',
-    Description: 'lambda function S3 bucket location'
-  };
-  parameters.Parameters.CodeS3Prefix = {
-    Type: 'String',
-    Description: 'lambda function S3 prefix location'
-  };
-  parameters.Parameters.GitSha = {
-    Type: 'String',
-    Description: 'Deploy Gitsha'
-  };
-
-  return parameters;
-}
-
 function buildSnsDestination(options) {
 
   let sns = {
@@ -383,7 +344,7 @@ module.exports.buildSnsDestination = buildSnsDestination;
 module.exports.buildRole = roles.buildRole;
 module.exports.buildServiceAlarms = alarms.buildServiceAlarms;
 module.exports.buildLambda = lambdaResource.buildLambda;
-module.exports.buildParameters = buildParameters;
+module.exports.buildParameters = cloudfrontParameters.buildParameters;
 module.exports.buildFunction = buildFunction;
 module.exports.compileFunction = compileFunction;
 module.exports.buildFunctionTemplate = buildFunctionTemplate;
